Loader: Move ctypes hooks into a module.
Given the latest changes, pyiboot01_bootstrap's globals are gone
(more precisely None) after the script has finished. But the hooked
classes still reference to them (including 'sys' and 'os'). I decided to
make this into a module instead of more complicated code like
mix-in-classes and lot of import statements.
